pyproject.toml
[tool.poetry]
name = "beer-analytics"
version = "1.0"
description = "Beer Analytics website project"
authors = ["Christian Scheb <me@christianscheb.de>"]
license = "THE BEER-WARE LICENSE"

[tool.poetry.dependencies]
python = "~3.10"
Django = "^4.2"
django-environ = "^0.10"
django-meta = "^2.1"
kaleido = "0.2.1"
numpy = "^1.24.2"
pandas = "^1.5.3"
plotly = "^5.14.1"
pybeerxml = "^2.1.2"
requests = "^2.28.2"
scipy = "^1.10.1"
translitcodec = "^0.7.0"
lxml = "^4.9.2"
django-structlog = "^6.0.0"
mysqlclient = "^2.2.0"
google-api-python-client = "^2.108.0"
openai = "^1.0.1"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
gunicorn = "^21.2.0"
elasticsearch = "^8.12.0"
elasticsearch-async = "^6.2.0"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-exporter-otlp = "^1.20.0"
opentelemetry-instrumentation-django = "^0.42b0"
opentelemetry-instrumentation-dbapi = "^0.42b0"

[tool.poetry.dev-dependencies]
black = "^23.3.0"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
